moved Bank class under common
