+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/bankdto/response/Banks.kt
@@ -1,5 +1,6 @@
 package com.thenewboston.data.dto.bankapi.bankdto.response
 
+import com.thenewboston.data.dto.bankapi.common.response.Bank
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -16,24 +17,4 @@ data class BankList(
     val previous: String? = null,
     @SerialName("results")
     val banks: List<Bank>
-)
-
-@Serializable
-data class Bank(
-    @SerialName("account_number")
-    val accountNumber: String,
-    @SerialName("ip_address")
-    val ipAddress: String,
-    @SerialName("node_identifier")
-    val nodeIdentifier: String,
-    @SerialName("port")
-    val port: Int? = null,
-    @SerialName("protocol")
-    val protocol: String,
-    @SerialName("version")
-    val version: String,
-    @SerialName("default_transaction_fee")
-    val defaultTransactionFee: Int,
-    @SerialName("trust")
-    val trust: Double
 )
