+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/common/response/Bank.kt
@@ -1,20 +1,24 @@
-package com.thenewboston.data.dto.bankapi.bankdto.response
+package com.thenewboston.data.dto.bankapi.common.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class BankTrustResponse(
+data class Bank(
     @SerialName("account_number")
     val accountNumber: String,
     @SerialName("ip_address")
     val ipAddress: String,
+    @SerialName("node_identifier")
+    val nodeIdentifier: String,
     @SerialName("port")
     val port: Int? = null,
     @SerialName("protocol")
     val protocol: String,
+    @SerialName("version")
+    val version: String,
     @SerialName("default_transaction_fee")
-    val defaultTransactionFee: Double,
+    val defaultTransactionFee: Int,
     @SerialName("trust")
     val trust: Double
 )
